+++ CHANGELOG.md
@@ -19,6 +19,10 @@
   - [BREAKING CHANGE] Rewrite `bract.core.config/apply-inducer`
   - [BREAKING CHANGE] Remove `bract.core.config/apply-inducer-by-name`
   - Add `bract.core.config/apply-inducer-by-key`
+- Overhaul induce function
+  - Drop function `bract.core.util/apply-inducer`
+  - Refactor `bract.core.util/induce` to swap arity - now it looks like `clojure.core/reduce`
+- [TODO] Replace default dev config file `config.dev.edn` with `config/dev.edn`
 - [BREAKING CHANGE] Rename `bract.core.inducer/run-inducers` to `run-config-inducers`
 - Introduce `bract.core.inducer/run-context-inducers` to run inducers from context
   - Context key definition `:bract.core/inducers`